feat: add CharactersService for character management including CRUD, detection, and avatar handling
